Refactored RenderWordTarget to not need angles

+++ WordTarget/WordTargetCore/WordTargetLayout.cs
@@ -31,9 +31,9 @@ namespace WordTargetCore
         public ReadOnlyCollection<double> FracsInCircle3 => fracsInCircle3.AsReadOnly();
         public ReadOnlyCollection<double> FracsInCircle2 => fracsInCircle2.AsReadOnly();
 
-        public double startingAngle2 { get; set; }
-        public double startingAngle3 { get; set; }
-        public double startingAngle4 { get; set; }
+        public int startingAngle2 { get; set; }
+        public int startingAngle3 { get; set; }
+        public int startingAngle4 { get; set; }
 
 
         public readonly List<List<string>> wordsInCircle;
@@ -48,9 +48,9 @@ namespace WordTargetCore
             fracsInCircle5.Add(Renderer.FracForWord(words[0], 5));
             wordsInCircle = new List<List<string>>() { null, null, wordsInCircle2, wordsInCircle3, wordsInCircle4, wordsInCircle5 };
             fracsInCircle = new List<List<double>>() { null, null, fracsInCircle2, fracsInCircle3, fracsInCircle4, fracsInCircle5 };
-            startingAngle2 = 80.0;
-            startingAngle3 = 0.0;
-            startingAngle4 = 5.0;
+            startingAngle2 = 80;
+            startingAngle3 = 0;
+            startingAngle4 = 5;
         }
 
         public void AssignWord(string word, int circle)
